Clean up PaymentCell component heirarchy
